streamlit/requirements.txt
openai==0.28.1
Pillow==10.2.0
Requests==2.31.0
streamlit==1.27.2
streamlit_option_menu==0.3.6
streamlit_text_rating==0.1.5
streamlit-authenticator